functions/requirements.txt
pypdf==4.0.1
pdf2image==1.17.0
Pillow==10.2.0 